fix: refactor getTermsOfUseContent
