+++ web/src/auth/LoginPage.js
@@ -30,7 +30,7 @@ import {CountDownInput} from "../common/CountDownInput";
 import SelectLanguageBox from "../SelectLanguageBox";
 import {CaptchaModal} from "../common/CaptchaModal";
 import RedirectForm from "../common/RedirectForm";
-import {getTermsOfUseContent, isAgreementRequired, renderAgreement, renderModal} from "./SignupPage";
+import {isAgreementRequired, renderAgreement, renderModal} from "./SignupPage";
 
 class LoginPage extends React.Component {
   constructor(props) {
@@ -127,7 +127,7 @@ class LoginPage extends React.Component {
           this.onUpdateApplication(application);
           this.setState({
             application: application,
-          }, () => getTermsOfUseContent(this.state.application.termsOfUse, res => {
+          }, () => Setting.getTermsOfUseContent(this.state.application.termsOfUse, res => {
             this.setState({termsOfUseContent: res});
           }));
         });
@@ -139,7 +139,7 @@ class LoginPage extends React.Component {
             this.setState({
               application: res.data,
               applicationName: res.data.name,
-            }, () => getTermsOfUseContent(this.state.application.termsOfUse, res => {
+            }, () => Setting.getTermsOfUseContent(this.state.application.termsOfUse, res => {
               this.setState({termsOfUseContent: res});
             }));
           } else {
